combined s3.py and run.py
